backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
openai==1.54.0
httpx==0.27.0
python-dateutil==2.8.2
